rename server/app.js to server/index.js
